requirements.txt
fastapi
uvicorn
twilio
openai>=1.0.0
python-multipart
python-dotenv
